Factor guard into own file
