+++ ui/src/app/app-routing.module.ts
@@ -1,14 +1,10 @@
-import {Injectable, NgModule} from '@angular/core';
-import {
-  ActivatedRouteSnapshot,
-  Resolve,
-  RouterModule,
-  Routes,
-} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
 
 import {HomeComponent} from './data-browser/home/home.component';
 import {SearchComponent} from './data-browser/search/search.component';
+
+import {CohortResolver} from './guards/cohort-resolver.guard';
 
 import {CohortEditComponent} from './views/cohort-edit/component';
 import {HomePageComponent} from './views/home-page/component';
@@ -17,27 +13,12 @@ import {WorkspaceEditComponent} from './views/workspace-edit/component';
 import {WorkspaceShareComponent} from './views/workspace-share/component';
 import {WorkspaceComponent} from './views/workspace/component';
 
-import {Cohort, CohortsService} from 'generated';
-
 /*
  * Both of these symbols (the resolver and the route list), which have no
  * relevance outside this module, MUST be exported. See
  * https://github.com/angular/angular-cli/issues/3707#issuecomment-332498738
  */
-@Injectable()
-export class CohortResolver implements Resolve<Cohort> {
-  constructor(private cohortAPI: CohortsService) {}
-
-  resolve(route: ActivatedRouteSnapshot): Observable<Cohort> {
-    const wsNamespace = route.paramMap.get('ns');
-    const wsID = route.paramMap.get('wsid');
-    const cohortID = +route.paramMap.get('cid');
-
-    return this.cohortAPI.getCohort(wsNamespace, wsID, cohortID);
-  }
-}
-
-export const routes: Routes = [
+const routes: Routes = [
   {
     path: '',
     component: HomePageComponent,
